\name{Internal Functions}
\alias{.binary.activematrix.check}
\alias{.build.button.function}
\alias{.build.command.argument}
\alias{.eval.command}
\alias{.find.frame}
\alias{.give.doublequote}
\alias{.is.binary.matrix}
\alias{.isISA}
\alias{.make.correct.frame}
\alias{.makeResultList}
\alias{.makeSuperbiclustResultList}
\alias{.isListofBiclustGUIresults}
\alias{.makesearchdata}
\alias{.order.button.frames}
\alias{.output.sparse.txt}
\alias{.rcmdr.warning}
\alias{.transform.vector2text}
\alias{.update.biclustering.object}
\alias{robust.fuse.support}
\alias{SearchMethodData}
\alias{Setwd}
\alias{as.ExprSet}
\alias{.bicare2biclust}
\alias{.fabia2biclust}
\alias{.putbefore}
\alias{.correctdataforresults}
\alias{.tobiclust}
\alias{.isbiclustGUIresult}
\alias{.tobiclust_transf}
\alias{.isBiclustresult}
\alias{.isa2biclust}
\alias{.checkplotgridpref}
\alias{greenred}
\alias{redgreen}
\alias{viridis}
\alias{.GetEnvBiclustGUI}
\alias{.AssignEnvBiclustGUI}
\alias{indexedBic}
\alias{BCDIAG_isa2biclust}


\title{List of Internal Functions}

\description{List of internal functions used in various functions in the package.}
%\usage{
%as.ExprSet(x)
%.bicare2biclust(x)
%.binary.activematrix.check()
%.build.button.function(function.command,arg.names,button_result,new.frames,save)
%.build.command.argument(current.frame,command)
%.correctdataforresult(result)
%.eval.command(x)
%.fabia2biclust(x,thresZ=0.5,thresL=NULL)
%.find.frame(x,frame.name)
%.give.doublequote(x)
%.isbiclustGUIresult(x)
%.isBiclustresult(x)
%.is.binary.matrix(x)
%.isISA(x)
%.make.correct.frame(title,border,window)
%.makeResultList()
%.makesearchdata()
%.order.button.frames(new.frames)
%.output.sparse.txt(X,file)
%.putbefore(colnames,pre)
%.rcmdr.warning(x)
%SearchMethodData(data,type)
%robust.fuse.support(robust.list,RowxNumber,NumberxCol)
%Setwd(x=TRUE)
%.tobiclust(x,fabia.thresZ=0.5,fabia.thresL=NULL)
%.tobiclust_transf(methodresult,thresZ="0.5",thresL="NULL")
%.transform.vector2text(x)
%.update.biclustering.object(object,where="all")
%
%}
%\arguments{
%\item{function.command}{ }
%\item{arg.names}{ }
%\item{button_result}{ }
%\item{new.frames}{ }
%\item{save}{ }
%\item{current.frame}{ }
%\item{command}{ }
%\item{x}{ }
%\item{frame.name}{ }
%\item{title}{ }
%\item{border}{ }
%\item{window}{ }
%\item{X}{ }
%\item{file}{ }
%\item{object}{ }
%\item{where}{ }
%}

\details{Not to be used by the users.}


\author{Ewoud De Troyer \email{ewoud.detroyer@uhasselt.be}}

