\name{Template Script Functions}
\alias{.add.frame}
\alias{.combine.rows}
\alias{.grid.matrix}
\alias{.initialize.grid.config}
\alias{.initialize.grid.rows}
\alias{.initialize.new.frames}
\alias{cluster_template}
\alias{newtool_template}

\title{Template Script Functions}

\description{List of internal functions used in the template scripts.}
%\usage{
%.add.frame(input="plotdiagTab",type,frame.name,argument.names="",arguments="",
%    initial.values=c(),title="",border=FALSE,entry.width="2",
%    argument.values=c(),argument.types=c(),from=c(),to=c(),by=c(),length=c(),
%    button.name="",button.function="",button.data="",arg.frames=c(),
%    button.otherarg="",button.biclust="",save=TRUE,show=TRUE, button.width="12",
%    button.data.transf="matrix", new.frames=new.frames)
%cluster_template(methodname="",methodfunction,methodhelp="",data.arg,
%    other.arg,methodseed=methodseed,grid.config=grid.config,
%    grid.rows=grid.rows,new.frames,superbiclust.comp,bcdiag.comp,
%    data.transf="matrix",data.discr=FALSE,data.bin=FALSE,extrabiclustplot=FALSE,
%    methodshow=TRUE,methodsave=TRUE,extrabiclustforisa=FALSE)
%.combine.rows(input,rows,title,border,grid.rows,grid.config)
%.grid.matrix(input,data,grid.config=grid.config,...)
%.initialize.grid.config()
%.initialize.grid.rows()
%.initialize.new.frames()
%newtool_template(toolname="",methodname="",toolhelp="",
%    grid.config=grid.config,grid.rows=grid.rows,new.frames)
%
%
%}
%\arguments{
%\item{input}{ }
%\item{type}{ }
%\item{frame.name}{ }
%\item{argument.names}{ }
%\item{arguments}{ }
%\item{initial.values}{ }
%\item{title}{ }
%\item{border}{ }
%\item{entry.width}{ }
%\item{argument.values}{ }
%\item{argument.types}{ }
%\item{from}{ }
%\item{to}{ }
%\item{by}{ }
%\item{length}{ }
%\item{button.name}{ }
%\item{button.function}{ }
%\item{button.data}{ }
%\item{arg.frames}{ }
%\item{button.otherarg}{ }
%\item{button.biclust}{ }
%\item{save}{ }
%\item{show}{ }
%\item{new.frames}{ }
%\item{methodname}{ }
%\item{methodfunction}{ }
%\item{methodhelp}{ }
%\item{data.arg}{ }
%\item{other.arg}{ }
%\item{methodseed}{ }
%\item{grid.config}{ }
%\item{grid.rows}{ }
%\item{superbiclust.comp}{ }
%\item{bcdiag.comp}{ }
%\item{data.matrix}{ }
%\item{data.discr}{ }
%\item{data.bin}{ }
%\item{extrabiclustplot}{ }
%\item{methodshow}{ }
%\item{extrabiclustforisa}{ }
%\item{rows}{ }
%\item{data}{ }
%\item{...}{ }
%\item{toolname}{ }
%\item{toolhelp}{ }
%
%}

\details{More information of the use of these functions can be found in the vignette. (Guide to New Implementations)}


\author{Ewoud De Troyer \email{ewoud.detroyer@uhasselt.be}}

